+++ pkg/service/backupexecutor/scan.go
@@ -55,7 +55,7 @@ func makeBackupConfig(
 
 	config.ParallelRead = backupPolicy.GetParallelOrDefault()
 	config.ParallelWrite = backupPolicy.GetParallelOrDefault()
-	config.FileLimit = backupPolicy.GetFileLimitOrDefault() * megabyte // lib expects limit in bytes.
+	config.FileLimit = uint64(backupPolicy.GetFileLimitOrDefault() * megabyte) // lib expects limit in bytes.
 	config.RecordsPerSecond = util.ValueOrZero(backupPolicy.RecordsPerSecond)
 	config.Bandwidth = util.ValueOrZero(backupPolicy.Bandwidth) * megabyte // lib expects file size in bytes.
